Fix compile issue with clang

+++ src/engine.cpp
@@ -801,7 +801,7 @@ void StartupOneEngine(Engine *e, const CalTime::YearMonthDay &aging_ymd, const C
 	 * Note: TTDP uses fixed 1922 */
 	CalTime::Date begin_random_date = CalTime::ConvertYMDToDate(_settings_game.game_creation.starting_year + 2, 0, 1);
 	if (_settings_game.vehicle.vehicle_intro_randomisation && ei->base_intro > begin_random_date) {
-		e->intro_date = ei->base_intro + CalTime::DateDelta{GB(r, 0, 9)};
+		e->intro_date = ei->base_intro + CalTime::DateDelta{(int)GB(r, 0, 9)};
 	} else {
 		e->intro_date = ei->base_intro;
 	}
